+++ x/mint/module/module.go
@@ -21,7 +21,7 @@ import (
 )
 
 // ConsensusVersion defines the current x/mint module consensus version.
-const ConsensusVersion = 2
+const ConsensusVersion = 4
 
 var (
 	_ module.AppModuleBasic = AppModule{} //nolint:exhaustruct
@@ -108,6 +108,30 @@ func (am AppModule) IsAppModule() {}
 func (am AppModule) RegisterServices(cfg module.Configurator) {
 	types.RegisterMsgServiceServer(cfg.MsgServer(), keeper.NewMsgServerImpl(am.keeper))
 	types.RegisterQueryServiceServer(cfg.QueryServer(), keeper.NewQueryServerImpl(am.keeper))
+
+	// we don't have any data to migrate, so the migration itself is a no-op,
+	// but it's good to print that we're doing a migration
+	err := cfg.RegisterMigration(types.ModuleName, 1, func(ctx sdk.Context) error {
+		ctx.Logger().Info(fmt.Sprintf("MIGRATING %s MODULE FROM VERSION 1 TO VERSION 2", types.ModuleName))
+		return nil
+	})
+	if err != nil {
+		panic(fmt.Sprintf("failed to migrate x/%s from version 1 to 2: %v", types.ModuleName, err))
+	}
+	err = cfg.RegisterMigration(types.ModuleName, 2, func(ctx sdk.Context) error {
+		ctx.Logger().Info(fmt.Sprintf("MIGRATING %s MODULE FROM VERSION 2 TO VERSION 3", types.ModuleName))
+		return nil
+	})
+	if err != nil {
+		panic(fmt.Sprintf("failed to migrate x/%s from version 2 to 3: %v", types.ModuleName, err))
+	}
+	err = cfg.RegisterMigration(types.ModuleName, 3, func(ctx sdk.Context) error {
+		ctx.Logger().Info(fmt.Sprintf("MIGRATING %s MODULE FROM VERSION 3 TO VERSION 4", types.ModuleName))
+		return nil
+	})
+	if err != nil {
+		panic(fmt.Sprintf("failed to migrate x/%s from version 3 to 4: %v", types.ModuleName, err))
+	}
 }
 
 // InitGenesis performs genesis initialization for the mint module. It returns